pyproject.toml
[project]
name = "voice-ai-orch"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # LiveKit Agent Framework
    "livekit-agents[openai,silero,turn-detector]~=1.3",
    "livekit-plugins-cartesia>=1.2.18",
    "livekit-plugins-deepgram>=1.2.18",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.2.18",
    "livekit-api>=1.0.0",
    # RAG - LlamaIndex + Pinecone + Google Embeddings + Cerebras LLM
    "llama-index>=0.14.7",
    "llama-index-embeddings-google-genai>=0.1.0",
    "llama-index-llms-cerebras>=0.1.0",
    "llama-index-vector-stores-pinecone>=0.4.0",
    "pinecone-client>=5.0.0",
    # Document parsing
    "pypdf>=6.1.3",
    # API Server
    "fastapi>=0.121.0",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.2.1",
]
